# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Nativos 3D Cutter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  createdAt: string;
  updatedAt: string;
  modelCount: number;
  operationCount: number;
}

export type ModelFormat = typeof ModelFormat[keyof typeof ModelFormat];


export const ModelFormat = {
  stl: 'stl',
  obj: 'obj',
  ply: 'ply',
  '3mf': '3mf',
} as const;

export interface Model {
  id: number;
  projectId: number;
  filename: string;
  format: ModelFormat;
  faceCount: number;
  vertexCount: number;
  /** @nullable */
  fileSizeBytes?: number | null;
  /** @nullable */
  volumeMm3?: number | null;
  /** @nullable */
  boundingBoxMm?: string | null;
  createdAt: string;
}

export type OperationType = typeof OperationType[keyof typeof OperationType];


export const OperationType = {
  cut: 'cut',
  selection: 'selection',
  repair: 'repair',
  export: 'export',
} as const;

export type OperationStatus = typeof OperationStatus[keyof typeof OperationStatus];


export const OperationStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Operation {
  id: number;
  modelId: number;
  projectId: number;
  type: OperationType;
  status: OperationStatus;
  /** @nullable */
  faceCount?: number | null;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface ProjectDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  createdAt: string;
  updatedAt: string;
  models: Model[];
  operations: Operation[];
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  thumbnail?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  thumbnail?: string;
}

export type ModelInputFormat = typeof ModelInputFormat[keyof typeof ModelInputFormat];


export const ModelInputFormat = {
  stl: 'stl',
  obj: 'obj',
  ply: 'ply',
  '3mf': '3mf',
} as const;

export interface ModelInput {
  projectId: number;
  filename: string;
  format: ModelInputFormat;
  faceCount: number;
  vertexCount: number;
  fileSizeBytes?: number;
  volumeMm3?: number;
  boundingBoxMm?: string;
}

export type OperationInputType = typeof OperationInputType[keyof typeof OperationInputType];


export const OperationInputType = {
  cut: 'cut',
  selection: 'selection',
  repair: 'repair',
  export: 'export',
} as const;

export interface OperationInput {
  modelId: number;
  projectId: number;
  type: OperationInputType;
  faceCount?: number;
  notes?: string;
}

export type OperationUpdateStatus = typeof OperationUpdateStatus[keyof typeof OperationUpdateStatus];


export const OperationUpdateStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface OperationUpdate {
  status?: OperationUpdateStatus;
  durationMs?: number;
  notes?: string;
  completedAt?: string;
}

export type ExportFormat = typeof ExportFormat[keyof typeof ExportFormat];


export const ExportFormat = {
  stl_binary: 'stl_binary',
  stl_ascii: 'stl_ascii',
  obj: 'obj',
  ply: 'ply',
  '3mf': '3mf',
} as const;

export interface Export {
  id: number;
  operationId: number;
  projectId: number;
  format: ExportFormat;
  filename: string;
  /** @nullable */
  fileSizeBytes?: number | null;
  createdAt: string;
}

export type ExportInputFormat = typeof ExportInputFormat[keyof typeof ExportInputFormat];


export const ExportInputFormat = {
  stl_binary: 'stl_binary',
  stl_ascii: 'stl_ascii',
  obj: 'obj',
  ply: 'ply',
  '3mf': '3mf',
} as const;

export interface ExportInput {
  operationId: number;
  projectId: number;
  format: ExportInputFormat;
  filename: string;
  fileSizeBytes?: number;
}

export type DashboardStatsOperationsByStatus = {[key: string]: number};

export type DashboardStatsExportsByFormat = {[key: string]: number};

export interface DashboardStats {
  totalProjects: number;
  totalModels: number;
  totalOperations: number;
  totalExports: number;
  totalFacesCut: number;
  recentProjectCount: number;
  operationsByStatus?: DashboardStatsOperationsByStatus;
  exportsByFormat?: DashboardStatsExportsByFormat;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  project_created: 'project_created',
  model_imported: 'model_imported',
  cut_completed: 'cut_completed',
  export_completed: 'export_completed',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  createdAt: string;
}

export type ListModelsParams = {
projectId?: number;
};

export type ListOperationsParams = {
modelId?: number;
projectId?: number;
};

export type ListExportsParams = {
operationId?: number;
projectId?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};