# lib/api-zod/src/generated/types/model.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Nativos 3D Cutter API
 * OpenAPI spec version: 0.1.0
 */
import type { ModelFormat } from './modelFormat';

export interface Model {
  id: number;
  projectId: number;
  filename: string;
  format: ModelFormat;
  faceCount: number;
  vertexCount: number;
  /** @nullable */
  fileSizeBytes?: number | null;
  /** @nullable */
  volumeMm3?: number | null;
  /** @nullable */
  boundingBoxMm?: string | null;
  createdAt: string;
}