# lib/api-zod/src/generated/types/modelInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Nativos 3D Cutter API
 * OpenAPI spec version: 0.1.0
 */
import type { ModelInputFormat } from './modelInputFormat';

export interface ModelInput {
  projectId: number;
  filename: string;
  format: ModelInputFormat;
  faceCount: number;
  vertexCount: number;
  fileSizeBytes?: number;
  volumeMm3?: number;
  boundingBoxMm?: string;
}