# lib/api-zod/src/generated/types/operation.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Nativos 3D Cutter API
 * OpenAPI spec version: 0.1.0
 */
import type { OperationStatus } from './operationStatus';
import type { OperationType } from './operationType';

export interface Operation {
  id: number;
  modelId: number;
  projectId: number;
  type: OperationType;
  status: OperationStatus;
  /** @nullable */
  faceCount?: number | null;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}