# lib/api-zod/src/generated/types/operationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Nativos 3D Cutter API
 * OpenAPI spec version: 0.1.0
 */
import type { OperationInputType } from './operationInputType';

export interface OperationInput {
  modelId: number;
  projectId: number;
  type: OperationInputType;
  faceCount?: number;
  notes?: string;
}